starknet: move 'signer' to 'core' directory
