+++ packages/starknet/lib/src/account.dart
@@ -11,11 +11,11 @@ import 'account/signer/base_account_signer.dart';
 import 'account/signer/stark_account_signer.dart';
 import 'contract/index.dart';
 import 'convert.dart';
+import 'core/signer/stark_signer.dart';
+import 'core/types/index.dart';
 import 'crypto/index.dart' as c;
 import 'presets/udc.g.dart';
-import 'signer/stark_signer.dart';
 import 'static_config.dart';
-import 'core/types/index.dart';
 
 enum AccountSupportedTxVersion {
   @Deprecated('Transaction version 0 will be removed with Starknet alpha v0.11')
