+++ packages/starknet/lib/src/argent/sessions.dart
@@ -1,10 +1,10 @@
 import 'package:freezed_annotation/freezed_annotation.dart';
 
+import '../core/signer/stark_signer.dart';
+import '../core/types/index.dart';
 import '../crypto/index.dart';
-import '../signer/stark_signer.dart';
 import '../snip/snip12/typed_data/shortstring.dart';
 import '../snip/snip12/typed_data/typed_data.dart';
-import '../core/types/index.dart';
 
 part 'sessions.freezed.dart';
 part 'sessions.g.dart';
