+++ packages/starknet/lib/src/core/index.dart
@@ -1,0 +1,2 @@
+export 'signer/index.dart';
+export 'types/index.dart';
