+++ packages/starknet/lib/src/core/signer/base_signer.dart
@@ -1,6 +1,4 @@
-// TODO: should be moved  to "core" package
-
-import '../core/types/felt.dart';
+import '../types/felt.dart';
 
 abstract class BaseSigner {
   Future<List<Felt>> sign(BigInt messageHash, BigInt? seed);
