+++ packages/starknet/lib/src/core/signer/index.dart
@@ -1,3 +1,2 @@
-// TODO: should be moved to "core" package
 export 'base_signer.dart';
 export 'stark_signer.dart';
