+++ packages/starknet/lib/starknet.dart
@@ -9,11 +9,10 @@ export 'src/argent/index.dart';
 export 'src/calldata.dart';
 export 'src/contract/index.dart';
 export 'src/convert.dart';
-export 'src/core/types/index.dart';
+export 'src/core/index.dart';
 export 'src/crypto/index.dart';
 export 'src/devnet/index.dart';
 export 'src/signer.dart';
-export 'src/signer/index.dart';
 export 'src/snip/index.dart';
 export 'src/static_config.dart';
 export 'src/util.dart';
